[rpc] throw correct Exception on local client based on the RemoteException type.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/32936

Closes https://github.com/pytorch/pytorch/issues/32732. Currently if a
UDF run in RPC throws an exception such as ValueError or TypeError, we wrap
this in a RemoteException on the callee side. When raising this on the caller
side, we currently raise a vanilla Exception. This diff changes it so that the
correct exception is thrown. Tested by changing the current rpc tests to assert
on the right type of error rather than just the base `Exception`.
ghstack-source-id: 97633548

Differential Revision: [D19700434](https://our.internmc.facebook.com/intern/diff/D19700434/)